cached.

+++ psmeta_individual_and_group_masks_afni.sh
@@ -42,7 +42,7 @@ regs='desc-confounds_timeseries'  # the token for fMRIPrep output nuisance regre
 anat='desc-preproc_T1w_brain'     # skull-stripped anatomical image
 deno='NR24a'                      # denoising strategy
 gmth=0.2                          # gray matter threshold between [0 1]
-tasks=("task-LocaVis1p75" "task-LocaAudio2p5" "task-AudioVisAssos1word")
+tasks=("task-LocaVis1p75" "task-AudioVisAssos1word" "task-AudioVisAssos2words")
 ## ---------------------------
 
 echo -e "========== START JOB at $(date) =========="
@@ -55,13 +55,19 @@ for subj in ${subjects[@]};do
   # prepare anatomical images
   gm_segment="$gdir/${subj}_${spac}_label-GM_probseg.nii.gz"         # gray matter segment in probability 
   gm_mask_t1="$sdir/${subj}_${spac}_mask-gm${gmth}_res-anat.nii.gz"  # gray matter mask in T1 resolution
-  cp -r $gm_segment $sdir  # copy gray matter image to individual masks folder
-  # create individual gray matter masks
-  3dcalc -a $gm_segment -expr "ispositive(a-$gmth)" -prefix $gm_mask_t1
+  if [ ! -f "$gm_segment" ];then
+    cp -r $gm_segment $sdir  # copy gray matter image to individual masks folder
+  fi
+  if [ ! -f "$gm_mask_t1" ];then
+    # create individual gray matter masks
+    3dcalc -a $gm_segment -expr "ispositive(a-$gmth)" -prefix $gm_mask_t1
+  fi
   # create functional masks and functionally constrained gray matter masks
   for task in ${tasks[@]};do
-    wdir="$adir/$subj/$task"             # the Working folder
-    oglm="${subj}_${task}_GLM.w${deno}"  # the token for the Output GLM
+    # setup path
+    wdir="$adir/$subj/$task"                       # the task Working folder
+    oglm="${subj}_${task}_GLM.wBIM.wPSC.w${deno}"  # the token for the Output GLM
+    # create F-based masks
     stats_afni="$wdir/$oglm/stats.${subj}_${task}+tlrc.[0]"               # AFNI statistics
     stats_mask="$sdir/${subj}_${spac}_mask-full-F_res-${task}.nii.gz"     # functional mask based on full F statistics
     gm_mask_bd="$sdir/${subj}_${spac}_mask-gm${gmth}_res-${task}.nii.gz"  # gray matter mask in task resolution
@@ -71,6 +77,11 @@ for subj in ${subjects[@]};do
     # create functionally constrained gray matter masks
     3dresample -master $stats_mask -prefix $gm_mask_bd -input $gm_mask_t1
     3dcalc -a $gm_mask_bd -b $stats_mask -expr 'a*b' -prefix $gm_fF_mask
+    # copy EPI-extent mask for partial coverage tasks
+    epi_mask="${subj}_${task}_${spac}_desc-brain_mask.nii.gz"
+    gm_epi_mask="$sdir/${subj}_${spac}_mask-gm-epi_res-${task}.nii.gz"
+    cp -r $wdir/$epi_mask $sdir
+    3dcalc -a $gm_mask_bd -b $sdir/$epi_mask -expr 'a*b' -prefix $gm_epi_mask  # gray matter mask constrained by EPI
   done
 done
 ## ---------------------------
@@ -80,20 +91,33 @@ if [ ! -d "$kdir/group" ];then
   mkdir -p $kdir/group
 fi
 # gray matter mask
-3dbucket -fbuc -aglueto $kdir/group/group_${spac}_label-GM_probseg.nii.gz $kdir/sub-*/sub-*_${spac}_label-GM_probseg.nii.gz
-3dTstat -prefix $kdir/group/group_${spac}_label-GM_probseg-mean.nii.gz -mean $kdir/group/group_${spac}_label-GM_probseg.nii.gz
-3dcalc -a $kdir/group/group_${spac}_label-GM_probseg-mean.nii.gz -expr "ispositive(a-$gmth)" \
-  -prefix $kdir/group/group_${spac}_mask-gm${gmth}_res-anat.nii.gz
- functional mask (i.e. full F) and functionally constrained gray matter mask
+ggm_segment="$kdir/group/group_${spac}_label-GM_probseg.nii.gz"
+ggm_segmean="$kdir/group/group_${spac}_label-GM_probseg-mean.nii.gz"
+ggm_mask_t1="$kdir/group/group_${spac}_mask-gm${gmth}_res-anat.nii.gz"
+if [ ! -f "$ggm_segmean" ];then
+  3dbucket -fbuc -aglueto $ggm_segment $kdir/sub-*/sub-*_${spac}_label-GM_probseg.nii.gz
+  3dTstat -prefix $ggm_segmean -mean $ggm_segment
+fi
+if [ ! -f "$ggm_mask_t1" ];then
+  3dcalc -a $ggm_segmean -expr "ispositive(a-$gmth)" -prefix $ggm_mask_t1
+fi
+# functional mask (i.e. full F) and functionally constrained gray matter mask
 for task in ${tasks[@]};do
-  3dmask_tool -input $kdir/sub-*/sub-*_${spac}_mask-full-F_res-${task}.nii.gz \
-    -prefix $kdir/group/group_${spac}_mask-full-F_res-${task}.nii.gz -frac 1.0
-  3dresample -master $kdir/group/group_${spac}_mask-full-F_res-${task}.nii.gz \
-    -prefix $kdir/group/group_${spac}_mask-gm${gmth}_res-${task}.nii.gz \
-    -input $kdir/group/group_${spac}_mask-gm${gmth}_res-anat.nii.gz
-  3dcalc -a $kdir/group/group_${spac}_mask-gm${gmth}_res-${task}.nii.gz \
-    -b $kdir/group/group_${spac}_mask-full-F_res-${task}.nii.gz -expr 'a*b' \
-    -prefix $kdir/group/group_${spac}_mask-gm-full-F_res-${task}.nii.gz  # manually check *mask-gm-full-F* masks to get *mask-gm-final* masks
+  gstats_mask="$kdir/group/group_${spac}_mask-full-F_res-${task}.nii.gz"
+  gepi_mask="$kdir/group/group_${task}_${spac}_desc-brain_mask.nii.gz"
+  ggm_mask_bd="$kdir/group/group_${spac}_mask-gm${gmth}_res-${task}.nii.gz"
+  ggm_fF_mask="$kdir/group/group_${spac}_mask-gm-full-F_res-${task}.nii.gz"
+  ggm_epi_mask="$kdir/group/group_${spac}_mask-gm-epi_res-${task}.nii.gz"
+  # group-averaged F mask
+  3dmask_tool -input $kdir/sub-*/sub-*_${spac}_mask-full-F_res-${task}.nii.gz -prefix $gstats_mask -frac 1.0
+  # group-averaged EPI mask
+  3dmask_tool -input $kdir/sub-*/sub-*_${task}_${spac}_desc-brain_mask.nii.gz -prefix $gepi_mask -frac 1.0
+  # resample GM to task resolution
+  3dresample -master $gstats_mask -prefix $ggm_mask_bd -input $ggm_mask_t1
+  # F constrained GM
+  3dcalc -a $ggm_mask_bd -b $gstats_mask -expr 'a*b' -prefix $ggm_fF_mask  # manually check *mask-gm-full-F* masks to get *mask-gm-full-F-final* masks
+  # EPI constrained GM
+  3dcalc -a $ggm_mask_bd -b $gepi_mask -expr 'a*b' -prefix $ggm_epi_mask  # manually check *mask-gm-epi* masks to get *mask-gm-epi-final* masks
 done
 ## ---------------------------
